Fix `--target_pattern_file` with Unicode labels
Target files specified via `--target_pattern_file` supported Unicode labels prior to Bazel 8 in some (unreliable) capacity, but this regressed due to the Unicode fixes in other parts of Bazel introduced in time for Bazel 8. This change ensures that the file is parsed in the "internal" string encoding, matching the rest of Bazel and restoring support for labels containing Unicode characters.