Avoid repo thrashing when switching between build and query
This was caused by `query` (and `mod`, for that matter) ignoring `--action_env` when computing the repository environment, which could result in repo rules rerunning if they had a dependency on a value set via `--action_env`.

This is fixed by honoring the value of `--action_env` for all commands that load, not just those that analyze.