Remove the builtin module `local_config_platform` when `--incompatible_disable_native_repo_rules` is set
This change requires rules_cc 0.1.1 and rules_java 8.12.0. If we can cherry-pick this back to 8.x, we could gracefully remove `local_config_platform` as part of a flag flip (instead of outright breaking the build of those who use it in 9.0).

Also removed duplicate error reporting on invalid repo names in `PackageLookupFunction` and `TargetPattern`, relying on the error reporting in `RepositoryDelegatorFunction` instead.

Closes #26167.

PiperOrigin-RevId: 764413610
Change-Id: Iea1525f6513600f5d79b4f3e302598aadee46c02