Initial support for header compilation direct deps
This adds a new output to header compilation actions which is used as the classpath for downstream header compilations, instead of using the same output for downstream header and regular compilations. Using a separate output allows trimming information that isn't needed by header compilation, but is required for regular compilation, including methods. This results in a smaller cache key for header compilation actions.

PiperOrigin-RevId: 764512852
Change-Id: Id1a16ae9d3f1328df273772ce7264fe271436667