Adjust wording for canonical repository name description
The wording of the initial sentence ends up feeling circular in how it
defines the term, which I found difficult to parse.  I adjusted the
wording of the apparent repository name description as well to match
the phrasing of the new wording.  I don't like ending in prepositions,
but also find that it feels more precise in this sort of case.

PiperOrigin-RevId: 766204818
Change-Id: I919302d5fbe1c4883bdad7d42d108af66c05422e