Deflake cc_builtin_tests
One of the shell scripts used in this test uses a function that does
```
echo "$1" | (grep -q "$2" ...)
```
In order to check for the presence of a particular symbol in the output of
`nm`. The use of `-q` (aka `--quiet`) causes the test to occasionally fail with
a SIGPIPE error:

> -q, --quiet, --silent
>   Quiet;  do  not  write  anything  to  standard output. **Exit immediately with zero
>   status if any match is found, even if an error was detected**. Also see the  -s  or
>   --no-messages option.

(emphasis added)

When `grep` finds a match and exits immediately, in some relatively rare cases
`echo` hasn't finished writing yet. When `echo` tries to finish writing to the
now-closed pipe, it gets interrupted by the "Broken pipe" signal, causing the
failure.

PiperOrigin-RevId: 766239187
Change-Id: I41e9543367e325e05226de61f8f9f1853e761c8f