Enable Skyframe restarts during action rewinding.
The new behavior is only enabled for analysis caching. This moves
checkIfActionLostInputTooManyTimes from before rewind planning to after
as it would otherwise overcount when there are Skyframe restarts. This
could increase the number of repeated rewind attempts by 1, but the
threshold is somewhat arbitrary anyway.

Enabling Skyframe restarts is possible in the first place because action
rewinding runs in a Skyframe thread and it is straightforward to separate
its stateless and stateful steps.

PiperOrigin-RevId: 766609610
Change-Id: I4bcaced4cc5156327cfd23481e17991405ad91a5