Optimize equals on starlark providers with defined fields (StarlarkInfoWithSchema)
Previous implementation was from StructImpl, which sorts field names, calls getValue for each field. The latter is implemented using bisection. So equals had complexity O(nlogn), which became apparent after replacing LibraryToLink with a provider.

This implementation is using just `table`. For a provider class, it's always the same length.

PiperOrigin-RevId: 767593582
Change-Id: I1ca2498fb342f94f8068c8e7eec36fab2ad97824