allow --experimental_remote_cache_ttl=server to request remote cache entries that live as long as the Bazel server
Under this flag, Bazel will retain remote execution records in its
persistent action cache for as long as the server runs.  To this end, we
introduce a new FileArtifactValue expirationTime sentinel value of -2.

Note that we bump the VERSION of CompactPersistentActionCache.  This
way, the right thing will happen if an earlier version of Bazel that
does not have the logic to selectively clear the expirationTime ==
-2 entries inherits a persistent action cache from a later version of
Bazel.

In order to see if your CompactPersistentActionCache has server-duration
entries, you can see if:

    bazel dump --action_cache | grep expirationTime=-2

finds anything.