Add executable to stateful runfiles if necessary
This is already being done in the stateless case and avoids a costly check that needs to flatten the runfiles tree.

In fact, the check was not sufficient as it also considered the executable to be present if it is the target of a custom symlink - but both test actions and downstream deps using location expansion assume that the executable is present in its default location. A new test verifies these cases.