[8.3.0] Use a persistent lockfile to speed up reproducible extensions (#26195)
Reproducible extensions are not included in the regular lockfile in
order to keep it small and less prone to merge conflicts. But
reproducible extensions can still take significant time to evaluate,
which currently happens on every server restart.

This commit makes it so that Bazel maintains an additional lockfile,
stored under the output base and not visible to the user. It only
records information that remain correct indefinitely, such as the
results of reproducible extensions as well as hashes of module files
from registries. See the comments on BazelLockFileValue for a detailed
comparison of the contents of the two lockfiles.

Fixes #24723

Closes #24757.

PiperOrigin-RevId: 764624358
Change-Id: Iad6596c8d7e77e048bada7f62c21781b5a2c7c78 
(cherry picked from commit c8f750e1bf0fe28e15eb988666bceeb94082846e)

Closes #26130