[8.3.0] Avoid repo thrashing when switching between build and query (#26258)
This was caused by `query` (and `mod`, for that matter) ignoring
`--action_env` when computing the repository environment, which could
result in repo rules rerunning if they had a dependency on a value set
via `--action_env`.

This is fixed by honoring the value of `--action_env` for all commands
that load, not just those that analyze.

Context:
https://bazelbuild.slack.com/archives/C014RARENH0/p1748337429642779

Closes #26177.

PiperOrigin-RevId: 765088399
Change-Id: I276fe83d83407c733be88df797ad8513ecb8ba35
(cherry picked from commit 9b90f4011dd669383cd96c79763163c49d137502)

Fixes #26178