Skycache: Handle zstd compressed bytes in the node deserializer
Without compression we have seen values of 39MB. With this and a follow up change that modifies the serializer we can bring those under 12MB (from 39MB to 4MB).

The compression version is prefixed by the bytes 0x8000 which can't appear as part of a var int in CodedOutputStream. Bytes from the node itself are compressed with Zstd.

We have tried other solutions while iterating on this problem, front compression among them which didn't help as much as Zstd compression.

The followup change that adds compression to node serialization will regress wall time on writer builds by about 35%. We need to evaluate the trade-offs down the line of storage savings vs. lag from writer builds in order to decide whether we would compress nodes selectively based on a threshold of their size.

RELNOTES:none
PiperOrigin-RevId: 770167140
Change-Id: Ib348f89e8c64894df70f0792e7e513762d9e5b48