Fix `GraphOutputWriter.partitionFactored()` to handle cycles robustly.
The bug lay in the fact that `graph.getRoots()` returns 0 nodes if the from/to targets lie on a cycle (when using `somepath()` or `allpaths()`), which means we are not processing any nodes at all. The fix is to just run `processSuccessors()` on successors of all nodes in the graph (we don't need to traverse in any particular order, i.e. BFS in this case).

PiperOrigin-RevId: 770476487
Change-Id: I433548908509b9dcb7e6f30821bb398168523c53