Implement CppModuleMapAction as a "simple" file write action in Starlark.
We can implement the required logic in a callback passed to "map_each"
that operates on a struct containing all the relevant data. This avoids
the need for multiple closures to handle the different cases.

We cannot preserve the mnemonic "CppModuleMap" due to the use of a
simple write action.

PiperOrigin-RevId: 770576415
Change-Id: I3ab9dc15f8c6806f2e73cdf80f805e0a67c2debe