Do not commit rapid suicide when the PID file has been deleted during shutdown.
The PidFileWatcher daemon thread will abruptly kill the server process
when it notices the pid file has been deleted (unknown commit).
But this file is deleted during shutdown in a shutdown hook!
This means the server might kill itself even if other shutdown hooks are still
running.

There's really no need for such a drastic over-reaction over the removal
of the pid file when it has been removed by the process itself, so we
add a new shutdown hook to cooperatively stop the watcher thread before
deleting the file it cares about.

This is particularly relevant if Blaze was built with code coverage
enabled: coverage processing occurs in a shutdown thread and may take
several seconds to complete.
PiperOrigin-RevId: 770711868
Change-Id: I0334070e1400747489d79404c2ab607b8e36b1ef