use DefaultInfo in Bazel's embedded Starlark
The Starlark embedded in Bazel, that can get used when building other repos, has problems exposed by enabling

    --incompatible_disable_target_default_provider_fields

Since this code is embedded in Bazel, there is no way for users to patch it in the course of trying to disable target default provider fields in their own repos.  We here address these issues.

Closes #26276.

PiperOrigin-RevId: 770991995
Change-Id: I4b29e9d52dd1789afe7de990ca9500812a8ec211