[8.3.0] Add executable to stateful runfiles if necessary (#26288)
This is already being done in the stateless case and avoids a costly
check that used to flatten the runfiles tree.

In fact, the check was not sufficient as it also considered the
executable to be present if it is the target of a custom symlink - but
both test actions and dependents using location expansion assume that
the executable is present in its default location. A new test verifies
these cases.

Fixes #26242

Closes #26246.

PiperOrigin-RevId: 770712680
Change-Id: I8eaf1185909c201c28109da9849407ede2e3844f 
(cherry picked from commit d60d578f330da7a78f6d237c6ac7bf447c6a5d2e)

Fixes #26250