[8.3.0] use DefaultInfo in Bazel's embedded Starlark (#26289)
The Starlark embedded in Bazel, that can get used when building other
repos, has problems exposed by enabling

    --incompatible_disable_target_default_provider_fields

Since this code is embedded in Bazel, there is no way for users to patch
it in the course of trying to disable target default provider fields in
their own repos. We here address these issues.

Closes #26276.

PiperOrigin-RevId: 770991995
Change-Id: I4b29e9d52dd1789afe7de990ca9500812a8ec211

Commit
https://github.com/bazelbuild/bazel/commit/784685969d11d096c5d4a65f11e1c3c6ee2427b8

Co-authored-by: David Sanderson <32687193+dws@users.noreply.github.com>