Skycache: Compress invalidation nodes above threshold
Threshold for compressing currently set at 580 bytes through experimentation. This threshold can be modified in the writer freely without having to modify the deserializer. This change also modifies the compression level which is set at the fastest now (-7). The compression level can also be tweaked without having to modify the deserializer.

On the example target used, compressing every node without a threshold yields a 30% reduction in storage for invalidation nodes. Adding the threshold yields 37%. This is because very small nodes actually use more bytes when compressed. In addition to that, without a threshold writer builds have a wall time regression but with the threshold I wasn't able to detect it.

There will be serialized nodes that have the prefix with the magic bytes 0x8000 and those without. If it has the magic bytes, it means it's compressed.

RELNOTES:none
PiperOrigin-RevId: 772120970
Change-Id: I57e1faf3a94e74fd77fcd95ae683788ef4dc1c95