Make coverage collection more resilient to test failure
`bazel coverage foo` runs the tests in target `foo` and collects
resulting code coverage. Note there may be arbitrarily many tests
defined in target `foo`.

Without the changes in this commit, if only one of the tests in `foo`
fails, it defeats coverage collection for all other tests in that
target.

This is fixed by deferring the `exit $TEST_STATUS` call to run after
`$LCOV_MERGER_CMD` is run.