Replace native LibraryToLink with a Starlark provider (StarlarkInfo)
LibraryToLink is created/produced only in create_library_to_link.bzl. (There are
3 ways to create them - through cc_common.link, .create_linking_context_from_compilation_context and .create_library_to_link).

LibraryToLink is consumed in Starlark production code and in native tests.

Rewrite native LibraryToLink into a deprecated wrapper for the Starlark provider.
Deprecated because it must only be used in tests and not production. This minimizes changes to
tests in a single change.

Delete previously used cc_internal.create_library_to_link

Implement a simple cc_internal.freeze. LibraryToLink is added to a depset by the users (because we have a second func in cc_common, that consumes a depset of them). depset requires it to be frozen.

There's a small memory regression coming with the change, due to worse memory structure of a starlark provider in comparison to Autovalue. A lot of the regression could be removed if Starlark providers knew about field's default value and wouldn't encode them on each instance.

PiperOrigin-RevId: 772873382
Change-Id: I42ff8df66576d229eb9a92548a3eadbed0ee1376