Remove symbolForEquality from LinkOptions
symbolGenerator was introduced because of action non-determinism b/120206809. Nowadays linking action is essentially a pure Starlark action, created out of native providers. This change makes/brings closer part of the providers to being essentially Starlark providers. Combined, pure Starlark actions and pure Starlark providers must not result in action non-determinism. If they do, we have a bigger problem.

The LinkOptions handled here are only user supplied linkoptions, that are passed in at the end of the command line. They are not start-end lib or library related flags passed to command line. The cl causes a possible reordering of user supplied link options. There are possible breakages because of Hyrum, but probably not a lot and I deem it's not worth running a TGP.

The cl (if it works) makes rewrite of those providers to Starlark a lot easier.

PiperOrigin-RevId: 773600517
Change-Id: I0765806817ddf446f0531f2efece7e99181e8d14