Don't store dependency cycles in SkyQueryEnvironment queries
For SkyQuery, nodes that are part of cycles are internally noted as containing a cycle but are otherwise included in results, no matter how many cycles the graph contains.

This was the case previously (cycles were logged with INFO but otherwise not noted to the user), but there used to be an arbitrary cap of 20 cycles in the graph; cycles beyond this limit of 20 would cause problematic nodes to never be marked finished, and would thus lead to problematic graph state with symptoms described in b/419741327.

This 20 cycle limit was instituted in CL/61295131 to prevent OOMs in cycle-intense graphs. The reasoning was that the full cycle was stored for each node in a cycle, which caused problems for graphs with lots of cycles. To mitigate this, we avoid storing cycle information for these nodes in this mode. (This information wasn't used in SkyQuery anyway, so this is no notable loss! This is verified by implementing a new subtype to CycleInfo that throws an error if details are accessed)

Note that this does not yet fix b/421176399; nodes that are part of cycles are still part of a stderr WARNING that the target doesn't exist, though this has no effect on the query results given to stdout. That fix would involve relaxing this log statement to avoid logging nodes as "missing" if they are only in error because of a cycle.

TEST: Manaully looked at a heap pprof to verify that CycleInfo is no longer a nontrivial portion of the heap in graphs with many cycles
TEST: Created a manual regression test with a manufcatured graph of 500 cycles of 10 nodes each
TEST: Manually verified to fix regression tests noted in b/419741327.
PiperOrigin-RevId: 773806623
Change-Id: If4e4bebe0a739e2114da8e106c5515cf55f853f4