Do not drop branches during coverage merging.
Coverage report merging should be consistent, no matter the order in
which reports are merged (the merge operation should be commutative and
associative). Dropping branches when the input reports disagree on the
number of branches violated this (we used to drop the RHS, so the
operation didn't commute).

The new merging strategy is simply to combine the execution values for
the same "(block, branch)" combination. If a combination doesn't exist
for one report, it's simply added to the merged result.

This may not still be "correct" for when the number of branches differs
between reports, but it's arguably "less wrong" since it's at least now
consistent.

Moving this logic to the "addBranch" method on SourceFileCoverage also
neatly handles the case where the same "block, branch" pair occurs more
than once in a single coverage report for a file. This can occur
legitimately when C++ templates are instantiated several times (see
https://github.com/bazelbuild/bazel/pull/26342#issuecomment-2992265793
and https://github.com/bazelbuild/bazel/issues/12142).

To simplify the merging logic and avoid having to handle "BA" and "BRDA"
style branches differently, we ensure that BranchCoverage always has a
branch identifier set. This will just be a simple incrementing counter
when parsing BA lines. A new test to ensure BA lines are correctly
printed out has been added to verify the change to BranchCoverage
doesn't break anything in that regard.

Fixes #12142

PiperOrigin-RevId: 775577636
Change-Id: I1c23c89ca9c4477aeee7f4a7cca70612248c8a1c