[8.4.0] Fix hang with force fetching + repo contents cache (#26413)
With the repo contents cache enabled, `bazel fetch --force` (and the
now-removed `bazel sync`) would cause Bazel to hang forever. The reason
is that writing into the repo contents cache causes a Skyframe restart
very late into `RepositoryDelegatorFunction`, and the force-fetch bit
causes the restarted function to consider it a cache miss, writing into
the repo contents cache again, causing an infinite loop.

This PR changes it so that a restarted `RepositoryDelegatorFunction`
will remember whether it had _just_ finished a fetch before; if so, it
will use the cache, even if force-fetch is on.

Fixes https://github.com/bazelbuild/bazel/issues/26389.

Closes #26409.

PiperOrigin-RevId: 776306095
Change-Id: I52ad7aeb6581a16268d26a8b142041dc0e748768

Commit
https://github.com/bazelbuild/bazel/commit/bf67251521eef5bd9effafd76f0c9739c86a355d

---------

Co-authored-by: Xdng Yng <wyverald@gmail.com>