Add more details for `--bes_upload_mode`
The previous documentation only mentioned the options, but did not provide guidance on what they do.

This is especially important for `fully_async`, as it's not immediately clear from the name that lifecycle events are almost always dropped without warning. In general, I think the options should be renamed for clarity as `nowait_for_upload_complete` also waits for uploads to complete, just at a different time.