Make `--guard_against_concurrent_changes` false positives less likely
If the proxy differs (which should be rare), we can compute a digest to know with certainty whether the input changed. Since a failed check doesn't just skip the cache upload but also emits a warning, we can avoid false positives in this way and ensure that the warning is actionable.