Remove caching of spawn input maps
There are only two consumers of the spawn input map that actually look at the keys: the runner that executes the spawn and the deprecated expanded spawn log. All other callers actually just want to walk the expanded inputs. It's thus not clear whether there is any benefit to caching the full map, which may be retained for an extended period of time through `RemoteAction` during async uploads.

Instead, this change introduces a lightweight way for walking the expanded inputs of a spawn as an `Iterable` and updates all other consumers to use it. The only expected regression is that the expanded spawn log now causes the input mapping to be computed twice, but it has very large overhead anyway.

Apart from an expected improvement to memory usage, the motivation for this change is to give runners more freedom in how they expand spawn inputs. In the future, this could be used to stage special artifacts as a whole rather than expanded into their individual children, which can help with sandbox performance.