Remove LineCoverage and use a simple map of line_no -> execution_count.
LineCoverage stored three things:

* Line number
* Execution count
* Optional checksum for the line

The first was wasteful; the line number was already known from the
map maintained by SourceFileCoverage.

The third is almost never used; it's produced by geninfo (which we don't
use) and optionally by coverage.py (disabled by default). Keeping it
around is expensive given how many lines may be included in combined
coverage reports and there does not appear to be any demand for it.

See https://github.com/bazelbuild/bazel/issues/26383 for motivation.
This is not a "fix" because the underlying issue remains, but it is a
minor improvement.

PiperOrigin-RevId: 779156319
Change-Id: I484c229165c4f993d6d87617ce8af0e6c108756e