Make `apple_crosstool_transition` fall back to `--platforms` instead of `--cpu`
This change modifies `apple_crosstool_transition` to fall back to the value of `--platforms` instead of `--cpu` if the value of the apple platform specific cpu flag is not set.

This change is guarded by `--use_platforms_in_apple_crosstool_transition`.

`//command_line_option:cpu` is removed from the transition outputs as it is specified for each platform via platform flags in unknown commit

PiperOrigin-RevId: 780647128
Change-Id: Ifcffa3cd8e37ead17fc696a74472cc6de6342291