[platforms] Allow targets to declare themselves incompatible with the target platform
This is useful for cases where compatibility depends on either an optional toolchain
or the content of a toolchain.

One example for when this would be used is for `{java,python}_library` which have a toolchain dependency on `cc_toolchain`: if all transitive sources are
Java/Python sources, then not having a `cc_toolchain` for the platform is perfectly fine. If the target has JNI dependencies, then the target can now mark
itself as incompatible if there is no `cc_toolchain` available.

Arguably, the example with the missing `cc_toolchain` isn't perfect as one could argue that the JNI dependency will fail to build first and the scenario
can't happen. I have a slightly more complex case internally where it depends on some fields within a toolchain whether the target is compatible with the
target/exec platform or not. Expressing this with multiple toolchains or constraints isn't feasible here as that would lead to 100s or 1000s of constraints.
This change provides a feasible solution to decide compatibility by the rule implementation itself.