Avoid `--guard_against_concurrent_changes` false positives
If the proxy differs (which should be rare), we can compute a digest to know with certainty whether the input changed. Since a failed check doesn't just skip the cache upload but also emits a warning, it's worth avoiding false positives in this way and keeping the warning actionable.

Sent in response to two Slack threads about `--guard_against_concurrent_changes` warnings:
* https://bazelbuild.slack.com/archives/CA31HN1T3/p1751448125096509
* https://bazelbuild.slack.com/archives/CA31HN1T3/p1751033926924599

Closes #26455.

PiperOrigin-RevId: 781066458
Change-Id: Idf87fa64955eb8c4634706b02dda15d38461ce39