Add executables() query function
This is useful for trying to come up with a list of targets that can be
run with `bazel run`. Transitionally to workaround the lack of this
folks would use `*_binary` to approximate this, but there are many
community rules that don't follow that pattern for things that run
scripts, etc.