Fix order-dependent test in ActionCacheCheckerTest
The test `testDifferentFiles` was failing because the `WriteEmptyOutputAction` was not overwriting the output file if it already existed. This caused the test to fail if a previous test had written to the same file.

This change removes the check for the existence of the file, ensuring that the file is always overwritten.

PiperOrigin-RevId: 781972788
Change-Id: Ie07886836b445c2b695a8bdfb896b41cb867af24