Always output BRDA lines in CoveragaeOutputGenerator by default.
Adds a flag to determine if the tool should output BA lines or BRDA
lines for branch coverage with the latter being the default.

This means we will no longer output "mixed" data and allows us to
simplify the handling and storing of branch information: The data in
BranchCoverage is now much simpler to interpret.

Further, BA lines are only used internally with the gcov tooling having
switched to BRDA lines long ago so it is quite unlikely that Bazel users
will ever notice a change.

PiperOrigin-RevId: 783253678
Change-Id: I5c11aef3dd654be644fbfc028c9efc023c1ff75a