Stop using the remote HTTP cache when it becomes unreachable
While Bazel is already able to fall back to local execution when the remoote cache becomes unreachable, it continues making calls to the remote cache for subsequent actions, which produces significant delays; this is especially bad in case of network timeouts (the default TCP timeout is 21s on Windows and 127s on Linux/Mac). Time is lost both when attempting to hit the cache and to store into it. This issue can be reproduced by specifying a non-existing address for the cache, such as `--remote_cache=https://10.255.255.1/`.

To address this issue, this PR makes the following changes to the circuit breaker:
- The retrier now distinguishes between success, transient error (which may be retried) and permanent error (which may not not be retried). Suitable classifiers are defined for gRPC and HTTP.
- A new `DEFAULT_MIN_FAIL_COUNT_TO_COMPUTE_FAILURE_RATE` threshold is introduced to reduce the time interval required to reach the threshold in case of a connection timeout. The value is chosen to be slightly more than the number of failures within the window interval with the default `--experimental_remote_failure_rate_threshold=10`.

Closes #24685.

PiperOrigin-RevId: 784172174
Change-Id: Ic16b62b781bc5fd9608effe35154836b1b3a557a