Skycache: Fix Skycache readers crashing with aliases
Builds where the alias itself is not in the active directories but the actual value is remotely cached resulted in a crash due to missing package data not stored by Skycache. This package data is used to get the actual value's target data. Remotely cached values store the target data separately anyway, so the solution is to use that directly instead of going through the package.

RELNOTES:none
PiperOrigin-RevId: 784188499
Change-Id: Idea95ff3d3a0338e62efbaaeeabdae0fe55f68a4