[8.4.0] Skip strategies incompatible with path mapping during resolution (#26563)
Spawn strategies that aren't sandboxed are now skipped while resolving a
strategy for a path mapped spawn, which allows non-sandboxed strategies
to be listed first and also results in a better error message if no
compatible strategy is available (instead of an obscure exec failure).

Worker execution is special in that it can always sandbox if needed, but
may not do so based on options. For this purpose, path mapped spawns are
now treated just like dynamically executed spawns by forcing sandboxing.

In the future, the same mechanism can potentially be used to run spawns
for rewound actions in Bazel, which must not delete or override their
previous outputs.

Work towards
https://github.com/bazelbuild/bazel/discussions/22658#discussioncomment-12431355
Fixes https://github.com/bazel-contrib/rules_go/issues/4366

Closes #25430.

PiperOrigin-RevId: 784040555
Change-Id: I26e110801ce84c41aef5f3e03e34879c595ebedf

Commit
https://github.com/bazelbuild/bazel/commit/a42c98185717661c5425aa4161289381f4811c9c

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>