Markdown syntax support for CLI docs
CLI reference docs contain a mix of HTML and markdown syntax, neither of which is supported. HTML is escaped and markdown is not processed at all.

To improve the readability of and establish a markup standard for documentation, markdown (CommonMark) support has been added for CLI help text. `org.commonmark:commonmark:0.25.0` is used for conversion to HTML.

Available syntax is restricted due to the existing HTML escape logic. The following is confirmed to work and covered by tests.

* Code spans
  ```markdown
  `code span`
  ```
* Code block
  ````markdown
  ```
  code block
  ```
  ````
* Emphasis
  ```markdown
  *emphasis*
  ```
* Strong emphasis
  ```markdown
  **strong emphasis**
  ```
* Inline link
  ```markdown
  [inline link](/url (title))
  ```
* Reference link
  ```markdown
  [reference link][ref]
  [ref]: /url (title)
  ```
* Hard line break
  ```markdown
  hard line\
  break
  ```
* Unordered list
  ```markdown
  - unordered
  - list
  ```
* Ordered list
  ```markdown
  1. ordered
  2. list
  ```
* Paragraph
  ```markdown
  paragraph 1

  paragraph 2
  ```

Resolves #26424

Closes #26432.

PiperOrigin-RevId: 784261954
Change-Id: I9fb12d574babe2ed9ade0a5e50f0e04d57077cca