Support lazy macro expansion for finalizer macros
To do this, we need to introduce a new abstraction - PackagePieces, with a
mutable implementation (EvalMacroFunction.RecursiveExpander) and an immutable
one (NonFinalizerPackagePiecesValue) - to hold a subset of a package's pieces and
to allow checking them for name conflicts.

(In a follow-up, I intend to also use EvalMacroFunction.RecursiveExpander to allow
assembling a full Package from its package pieces.)

We take care not to expand finalizers if non-finalizer pieces contain errors or if
there are name conflicts between them.

One wart is that the order of rules in native.existing_rules() is different
depending on whether we are in lazy or eager expansion mode: we don't preserve the
declaration order of rules or macros in an evaluated packageoid, and therefore,
under lazy macro expansion, the only reasonable thing to do with the keys of
native.existing_rules() is to present them sorted.

Working towards https://github.com/bazelbuild/bazel/issues/23852 and https://github.com/bazelbuild/bazel/issues/25539

PiperOrigin-RevId: 784280703
Change-Id: I72c9a1a14503a2f060f0349747c3527c06d889d3