Rewrite CcLinkingContext and ExtraLinkTimeLibraries to Starlark
ExtraLinkTimeLibraries and StarlarkDefinedLinkTimeLibraries are both contained in create_extra_link_time_library.bzl. It exposes create_extra_link_time_libraries and merge_~, which is used when creating and merging CcLinkingContext.

To reduce memory footprint 2 interface changes were needed:
- `linking_context.extra_link_time_libraries().extra_libraries() ==> linking_context._extra_link_time_libraries.libraries`
- `X.build_libraries(**kwargs) ==> cc_common.build_extra_link_time_libraries(X.libraries, **kwargs)
`

Code outside of Bazel was prepared for that in unknown commit

The whole functionality of CcLinkingContext is moved into cc_info.bzl. I believe that is a good location, since there's relatively little code related to CcInfo. Basically just merging it.

The rest of the change is just wiring everything together.

PiperOrigin-RevId: 784560526
Change-Id: I89d898eafcbb771a16a59cbab54d01e352c4ee03