Do not use directExecutor() for FileOpNodeCollector.onSuccess().
This speeds up the serialization of Skyframe data by a factor of ~3x.

It also makes the CPU load much larger in the interim, indicating good parallelism. In addition, the daemon serving the underlying FUSE file system also gets significant CPU time, indicating that the bottleneck may not be Bazel parallelism anymore; this is not certain, since the CPU use of Bazel falls off before the end of the command faster than expected. I haven't investigated further.

My vague understanding of this change is that when the ~last leaf FileOpNodeCollector completes, it propagates its results up the graph of FileOpNodeCollectors, which can correspond to a large part of the Skyframe graph. If one uses directExecutor(), this all happens in one thread. ForkJoinPool.commonPool() enables reasonable parallelism.

RELNOTES: None.
PiperOrigin-RevId: 785970184
Change-Id: Ie0699d1462241f9daf2b5c655b4e3cc43d6b272e