Send ACS invalidation lookups only if a different client id is used.
Avoid unnecessary invalidation of the deserialized keys if the client ID that was used in the previous invocation is the same as the current invocation. Otherwise, should Skycache lose cache state, ACS will unnecessarily cause cache misses on nodes that have been already deserialized locally, causing Bazel to invalidate them.

PiperOrigin-RevId: 786173869
Change-Id: Ia1ac71d6042012fc10928b0837e5d1f4a366ee93