Maintain Jacoco's order for branches.
This is preparation for an update to Jacoco in which its filters will
become sensitive to branch order for the branches on a line,
particularly for "branch replacement".

This change alters the visitor to record branch indices correctly where
before we were somewhat sloppy (since we didn't use that information).

We use this new branch index information to create "BranchExp" instances
of the correct size whenever we need to (populated with "null
expression" objects) during the mapping of instructions to expressions.
We can then ensure we populate BranchExp data in a manner that is
consistent with Jacoco's own notion of branch indices.

This will simplify things for us when we update the version of Jacoco we
use.

Jacoco's ordering of branches is not necessarily what might be
intuitively expected: given a conditional jump to a target, the "first"
branch is actually the fall-through, with the second branch being the
continuation at the target.
For switch statements, it's the case that the first branch is always the
"default" case.

As a result, there may be an observable change in the order of branches
in the final coverage output files for Java code. However, for branches
that have been "replaced" (in some switch statements for instance), this
will not occur since the new branches will continue to be output in
"walk" order instead of Jacoco's "logical" order. In any case, we have
never made any promises with respect to branch order so this should not
matter.

PiperOrigin-RevId: 786290804
Change-Id: I38f516fc305985522032b2e9770f46a5fa0165c3