Darwin system network stats: place calls to sysctl() in a loop
Calls to sysctl that extract dynamically sized buffers should typically be placed in a loop, because there is no guarantee that the buffer returned by the second call to sysctl() fits within the buffer size that was obtained during the first call.

Issues like these are fairly common. For example, I ran into a similar problem with golang.org/x/sys a couple of years ago:

https://github.com/golang/sys/commit/70f4e408de566fcb64f1c8649fe05a22958255be

Closes #26599.

PiperOrigin-RevId: 786299694
Change-Id: Iecb7b64c2b751732dd7eb2026c0e6154d13f380a