[8.4.0] Print interactive debug instructions in HardlinkedSandboxedSpawn (#26619)
When running with --experimental_use_hermetic_linux_sandbox, this allows
us to see the interactive debug instructions like we'd see without it.

Note that `EnterWorkingDirectory()` in `linux-sandbox` requires that the
working directory is under the sandbox_root path, so we need to include
`cwd=sandboxExecRoot()` when calling `describeCommand`. This is a
stricter requirement than for SymlinkedSandboxedSpawn which was changed
in #26448 more for UX reasons.

Fixes #26580.

Closes #26581.

PiperOrigin-RevId: 786165593
Change-Id: Ib1a2dfccb7731691314d88d10e4c6a168977c875

Commit
https://github.com/bazelbuild/bazel/commit/289e3859aedab2af7253b6053f6fd575bfd0f3af

---------

Co-authored-by: Wade Carpenter <wade@arista.com>