Don't fail when the same aspect is added more than once to AspectsList.
This will unblock Bazel erroring when the same aspect is passed twice on the command line, like `--aspects=foo.bzl%foo_aspect --aspects=foo.bzl%foo_aspect`.

This is common if the user specifies some aspect in a blazerc managed behind a `--config` flag, and passing the same `--config` flag on the command line.

PiperOrigin-RevId: 787009837
Change-Id: Ifba2807a9787493934d5fb4617fde60433138789