Support flag aliases on `platform(flags = [...])`.
When we have `--flag_alias=foo=//bar` and see `--foo=1`, we want to map that to `--//bar=1` as early as possible in options parsing. Options can be parsed in these places:

1. At the command line.
2. In `config_setting`
3. In transitions
4. On `platform()` `flag =` attributes.

This adds support for 4. Note that uses `ParsedFlagsFunction`, which is a generic flag parsing API any Bazel code can call. I believe there are a few other users of that function: scoped flags and platform mappings. All users automatically benefit from this change.

**Motivation**: removing native flag definitions when they're Starlarkified. APIs that still use `--native_flag_syntax` will continue working.

PiperOrigin-RevId: 788117433
Change-Id: Ie350669bc0e2eb7a5f879498488f5179429c805b