skycache incrementality/invalidation fix: correctly handle synchronization between state of deserialized nodes in the Skyframe graph and `RemoteAnalysisCachingState#deserializedKeys()` by streaming updates instead of a batch update.
For cache hits:

Record the deserialized `SkyKey` as soon as the cache hit happens in `SkyframeExecutor#remoteAnalysisCachingState` **and** before setting the value in the `NodeEntry`, instead of waiting until the end of the build for the batch update.

Specifically, we conservatively record a cache hit into `deserializedKeys` before the Skyframe graph is mutated. This prevents the case where a deserialized key is committed into the Skyframe graph, but `deserializedKeys` doesn't contain it due to an interrupt before the batch update at the end of the build, causing subsequent invalidations to be incorrect.

For cache misses/invalidated nodes:

We were batch deleting invalidated nodes from `deserializedKeys` too early, because `MemoizingEvaluator#delete()` doesn't delete the keys until the next evaluation begins. If an interrupt happens between the `deserializedKeys` deletion and the Skyframe deletion, then subsequent invalidations will be incorrect. To fix that, remove the keys only *after* `SkyframeProgressReceiver` has been notified by `EagerInvalidator` that the deletion is complete for that specific key.

PiperOrigin-RevId: 788470078
Change-Id: I5790f53a22a7887a6f8581af0fc3208a9efe2f94