Optimize `RemoteOutputChecker.addOutputToDownload`
With `--remote_download_toplevel`, `RemoteOutputChecker#addOutputToDownload` is transitively called by `BuildDriverFunction`, which unconditionally runs for each top-level target on each (even ultimately no-op) build. It's applied to each top-level output as well as runfiles, making it a very hot code path.

This change exploits the fact that artifact paths are never prefixes of each other to simplify and speed up the data structure that maintains the paths to download with BwoB. Instead of a trie built out of nested `ConcurrentHashMap`s, it uses a single `ConcurrentSkipListSet` to look up paths via `floor`. This avoids all string operations as well as a number of hash maps lookup linear in the number of segments of a path.

The following benchmark results show the effect of the change on the Bazel repo itself (disabling as much I/O as possible to make the result less noisy):

```
$ hyperfine --warmup 5 --runs 20 './bazel-before --nosystem_rc --nohome_rc build --nobuild_runfile_links --noexperimental_check_output_files --incompatible_strict_action_env --disk_cache=/Users/fmeum/bazel-disk //src/test/...'

  Time (mean ± σ):     512.4 ms ±  50.7 ms    [User: 13.6 ms, System: 12.3 ms]
  Range (min … max):   447.0 ms … 643.5 ms    20 runs

$ hyperfine --warmup 5 --runs 20 './bazel-after --nosystem_rc --nohome_rc build --nobuild_runfile_links --noexperimental_check_output_files --incompatible_strict_action_env --disk_cache=/Users/fmeum/bazel-disk //src/test/...'

  Time (mean ± σ):     303.6 ms ±  36.9 ms    [User: 14.3 ms, System: 13.4 ms]
  Range (min … max):   270.0 ms … 383.6 ms    20 runs
```

Closes #26639.

PiperOrigin-RevId: 788514662
Change-Id: Ida0fc99bc7ab741418189e04b9fd98ad944dc2b6

Fix comparator used in ConcurrentArtifactPathTrie

The default comparator for `PathFragment` may sort unrelated paths between a path and a child path, which results in `contains` behaving incorrectly as soon as paths such as `foo/bar` and `foo/bar-baz` are present in the set. This is a bug introduced by #26639.

This is fixed by introducing a new `Comparator` for `PathFragment` that sorts `/` lower than any other character. This will also be used in future work on #21378 and is thus placed in `PathFragment`. It may become the default in the future, but not until Blaze has been confirmed to not rely on the order.

Closes #26714.

PiperOrigin-RevId: 792519419
Change-Id: I8bf8fc6b15cc5bcf0d303057a828fbd644032bab

Closes #26640